requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
passlib[bcrypt]
python-jose
python-multipart
alembic
python-dotenv
PyJWT
pydantic
